requirements.txt
fastapi
uvicorn
python-multipart
requests
transformers
pillow
opencv-python-headless
scikit-learn
pandas
pytest
httpx
xgboost
sqlalchemy
pymysql
onnxmltools
onnxconverter-common
python-dotenv
pandas-stubs
onnxruntime
scipy
sentencepiece
soundfile
datasets
fastapi
uvicorn
langchain
langchain-huggingface
transformers
torch
accelerate
matplotlib
seaborn
flake8
Black